arqmariav3_enhanced/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
MarkupSafe==2.1.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
google-api-python-client==2.108.0
beautifulsoup4==4.12.2
lxml==4.9.3
supabase==2.0.2
psycopg2-binary==2.9.7